==> 01-training-profiles-visualization.ipynb <==
import numpy as np
import os
import torch
import matplotlib.pyplot as plt
import seaborn as sb
import pandas as pd
import umap
import tqdm

feature_path = ''
df_path = ''

output_folder = 'output'
output_file = 'UMAP_of_Combined_CP.csv'

files = [x for x in os.listdir(feature_path) if x.startswith("features")]

features = []
idx = []
for f in tqdm.tqdm(files):
    feat, ids = torch.load(path + f)
    features.append(feat.numpy())
    idx.append(ids)

features = np.concatenate(features)

idx = np.concatenate(idx)

df = pd.read_csv(df_path)

reducer = umap.UMAP()
embedding = reducer.fit_transform(features)

df = df.iloc[idx]
df["X"] = embedding[:,0]
df["Y"] = embedding[:,1]

if not os.path.isdir(output_folder):
    os.mkdir(output_folder)
    
df.to_csv(f'{output_folder}/{output_file}',index=False)

plt.figure(figsize=(10,10))
sb.scatterplot(data=df, x="X", y="Y", hue="Collection", linewidth=0, s=5)
